Invert dependency between svg_file.py and svg_primitives.py
