+++ test/test_svg.py
@@ -2,7 +2,8 @@
 from textwrap import dedent
 
 from svg_timeline.svg_style_defaults import Color, ColorPalette
-from svg_timeline.svg import SVG, SvgElement, CascadeStyleSheet
+from svg_timeline.svg_file import SvgFile, CascadeStyleSheet
+from svg_timeline.svg_primitives import SvgElement
 
 
 def test_svg_element_getters():
@@ -34,7 +35,7 @@ def test_svg_element_str():
 
 
 def test_svg_header():
-    svg = SVG(width=800, height=600)
+    svg = SvgFile(width=800, height=600)
     header = dedent('''\
     <?xml version="1.0" encoding="UTF-8"?>
     <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"
@@ -48,8 +49,8 @@ def test_svg_defs_section():
         SvgElement('a', {'a1': 'hello', 'a2': 'world'}, 'asdf'),
         SvgElement('b', {'b1': 'hello', 'b2': 'world'}),
     ]
-    svg_with = SVG(width=800, height=600, definitions=definitions)
-    svg_without = SVG(width=800, height=600)
+    svg_with = SvgFile(width=800, height=600, definitions=definitions)
+    svg_without = SvgFile(width=800, height=600)
     defs_section = dedent('''\
     <defs>
       <a a1="hello" a2="world">asdf</a>
@@ -65,8 +66,8 @@ def test_svg_element_section():
         SvgElement('a', {'a1': 'hello', 'a2': 'world'}, 'asdf'),
         SvgElement('b', {'b1': 'hello', 'b2': 'world'}),
     ]
-    svg_with = SVG(width=800, height=600, elements=elements)
-    svg_without = SVG(width=800, height=600)
+    svg_with = SvgFile(width=800, height=600, elements=elements)
+    svg_without = SvgFile(width=800, height=600)
     element_section = dedent('''\
     <a a1="hello" a2="world">asdf</a>
     <b b1="hello" b2="world" />
@@ -76,7 +77,7 @@ def test_svg_element_section():
 
 
 def test_svg_footer():
-    svg = SVG(width=800, height=600)
+    svg = SvgFile(width=800, height=600)
     footer = dedent('''\
     </svg>
     ''')
